requirements_dev.txt
pip>=21.3.1
setuptools>=60.5.0
bump2version>=1.0.1
wheel>=0.37.1
watchdog>=2.1.6
flake8>=4.0.1
tox>=3.24.5
coverage>=6.2
Sphinx>=4.3.2
twine>=3.7.1

bandit>=1.7.1
black>=21.12b0
isort>=5.10.1
mypy>=0.930
mypy-extensions>=0.4.3
pre-commit>=2.17.0
pytest>=6.2.5
pytest-cov>=3.0.0

antspyx>=0.3.1; python_version <= "3.10"
matplotlib>=3.5.1
nibabel>=3.2.1
numpy>=1.22.0
pillow>=9.0.0
pydicom>=2.2.2
pymedio>=0.2.14
scikit-fuzzy>=0.4.2
scikit-image>=0.19.1
scikit-learn>=1.0.2
scipy>=1.7.3
SimpleITK>=2.1.1
statsmodels>=0.13.1
